backend/requirements.txt
boto3==1.35.78
flake8==7.1.1
black==24.10.0
Flask==3.1.0
gunicorn==23.0.0
flask-cors==5.0.0
argon2-cffi==23.1.0
pyjwt==2.10.1
